requirements.txt
biopython==1.76
certifi==2019.11.28
chardet==3.0.4
fire==0.2.1
idna==2.8
numpy==1.18.1
requests==2.22.0
six==1.14.0
squidbio==0.1
termcolor==1.1.0
urllib3==1.25.8
